fix: TODO.md audit — 7 P0 bugs, P1 batch, review findings
P0:
- weather: WMO snow/rain branch order (71-77/85-86 snow, 82 rain, 95+
  lightning; dead code removed) + zero-coordinate geocode (== null checks)
- relative time: formatAge floors totalSeconds (no more '42.738193s');
  single refcounted 60s ticker via useSyncExternalStore (one interval
  per page, not per item)
- search: kbd + title render the configured shortcut, not hardcoded S
- config: span .int() + integer sweep across 12 fields — with the review
  catch: collapse-after/collapse-after-rows min(-1) keeps glance's
  'disable collapsing' -1 semantic, limit min(0) keeps glance-parity 0
- server: TtlCache.clear() on successful config reload (widgets refetch
  after config edits); serveDist Cache-Control (immutable hashed assets,
  no-cache index/sw/registerSW/favicon/icon, short fonts) + malformed
  percent-encoding falls through to SPA instead of 500
- useConfig: failed /api/config fetch is retried on next mount (cache
  cleared on catch), happy-path cache kept
P1:
- unknown slug redirects home (<Navigate to='/' replace>)
- JetBrainsMono-SemiBold bundled; 600 weight no longer faux-bold
- duplicate manualChunks comment removed
P3: registry comment documenting intentional defense-in-depth re-validation

322/322 tests, tsc clean, doctor 100/100

+++ src/shared/config.ts
@@ -7,7 +7,7 @@ export const ColumnSchema = z.object({
   size: z.enum(['small', 'full']),
   widgets: z.array(WidgetSchema),
   // NEW — auto tiling only; number of grid tracks this tile spans.
-  span: z.number().min(1).max(4).optional(),
+  span: z.number().int().min(1).max(4).optional(),
 });
 export type Column = z.infer<typeof ColumnSchema>;
 
@@ -24,7 +24,7 @@ export const PageSchema = z.object({
   // grid tiles; 'collage' = dense bento grid + JS-measured row spans.
   tiling: z.enum(['columns', 'auto', 'collage']).optional(),
   // NEW — auto mode only; minimum tile width for auto-fit in px, default 300.
-  'min-column-width': z.number().min(1).optional(),
+  'min-column-width': z.number().int().min(1).optional(),
   columns: z.array(ColumnSchema).min(1).max(3),
 });
 export type Page = z.infer<typeof PageSchema>;
